fix: separate client hooks from analytics barrel to fix server/client boundary
BREAKING CHANGE: React hooks (useAnalytics, usePageView) are no longer exported from @/lib/analytics barrel

- Removed hook exports from src/lib/analytics/index.ts to prevent "Client Component cannot be imported into a Server Component" errors
- Updated all component imports to use direct hook paths: @/lib/analytics/hooks/use-analytics
- Updated documentation (CLAUDE.md and docs/analytics/) to reflect new import patterns
- The barrel (@/lib/analytics) now only exports server-safe utilities and services
- For React hooks, import directly from: @/lib/analytics/hooks/use-analytics or @/lib/analytics/hooks/use-page-view
- For server-side code, continue using: @/lib/analytics/server

This maintains proper server/client boundaries in Next.js 15 and prevents build errors when server modules import from the analytics barrel.
